Smaller final docker image
Split the build process into 2 parts: the first part does the build in
a docker image with the full haskell toolchain, etc. The second part
copies the necessary files for actually running (a statically compiled
binary & the HLint source) into a much smaller docker image.

The engine source had to be adjusted for this a bit: HLint reads its own
source code for default configuration, and we need to change the path it
looks in.